requirements.txt
Flask>=2.1.1
Flask-SQLAlchemy>=2.5.1
Flask-Login>=0.5.0
Flask-WTF>=1.0.1
Flask-Bcrypt>=1.0.1
WTForms>=3.0.0
email-validator==1.1.3
python-dotenv
mysql-connector-python

